test_heavy_vlms/requirements.txt
# test_heavy_vlms/requirements.txt
transformers>=4.45.0
torch
torchvision
accelerate>=0.26.0
bitsandbytes
protobuf
pillow
sentencepiece
av
librosa
# For Silero VAD
torchaudio
# For InternVL
einops
timm
# For Qwen-VL
tiktoken
transformers_stream_generator
# For Video-LLaMA
decord
# Utility and API
google-genai
google-generativeai
numpy
python-dotenv
opencv-python-headless
